Refactor compact-genome into interface and implementation.
Closes #54
