+++ implementation/compact-genome/src/implementation/mod.rs
@@ -1,0 +1,1 @@
+pub mod vector_genome_impl;
