+++ implementation/compact-genome/src/interface.rs
